// 10.front/myCocktails-ng/src/app/model/material/materialModel.ts
export interface MaterialModel { 
    /**
     * Material id
     */
    materialId?: number;
    /**
     * Material name.
     */
    materialName?: string;
    /**
     * Material category id.
     */
    categoryId?: number;
    /**
     * Material category name.
     */
    categoryName?: string;
}