// 10.front/myCocktails-ng/src/app/model/material/userMaterialModel.ts
/**
 * Materials API
 * API to manage material info.
 *
 * The version of the OpenAPI document: 0.2.10
 * Contact: okarians.302.dev@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * User material info model.
 */
export interface UserMaterialModel { 
    /**
     * User id.
     */
    userId?: string;
    /**
     * Material id list.
     */
    materialIdList?: Array<number>;
}